data: overflow guards, currency-aware money, MCC accuracy, jitter floor
retry.go switches from full jitter (which legitimately returns 0
delay) to equal jitter d/2 + rand(d/2) with an absolute 50ms floor.
Without this, a 5xx storm could translate into a tight instant-retry
loop against the bank.

money/Money grows overflow guards:
- Add and Sub use math/bits.Add64/Sub64 with explicit signed-overflow
  detection and return ErrOverflow on the boundary cases.
- Mul becomes (Money, error) — BREAKING but the only correct option:
  silent wrap on MaxInt64*2 was the worst possible behaviour for a
  banking SDK.
- Scale godoc warns about float64 precision past 2^53 and points
  callers to integer-only Mul for accounting-grade aggregation.

currency package: 10^Decimals via the new MinorPerMajor helper. JPY
and KRW pinned to 0 decimals, BHD/JOD/KWD/OMR/TND to 3, with their
alpha-3 mappings — so future code that touches these will work.

business/Account.BalanceMoney and BalancePoint.Money now multiply by
the currency's MinorPerMajor instead of hardcoding 100, so JPY (0
decimals) and the 3-decimal Gulf currencies round correctly.

acquiring.TipsInfo.Amount: int → int64. Mono returns minor units
that can in theory overflow 32-bit int on a 32-bit build (and Go
arm/386 is still in CI for many users).

mcc: split the 3000-3999 catch-all (was: Transport) into
  - 3000-3299 airlines, Transport
  - 3300-3499 car rental, Transport
  - 3500-3999 lodging, Hotels
and added a healthMCC override table for pharmacies, opticians,
doctors and hospitals (5912, 5122, 5975, 5976, 8011, 8021, 8031,
8041-8050, 8062, 8071, 8099). Existing TestCode_Category updated
for the lodging split; new regression tests for both directions.

bank.Transaction.MCCCode and OriginalMCCCode now sanitise via
validMCC (1..9999) before returning, so a wire glitch on the signed
int32 surfaces as CategoryUnknown instead of nonsense.

bank/Convert: added TestConvert_buySellSemantics with realistic Mono
quotes (RateBuy=41.50, RateSell=42.00) covering both directions to
lock in the buy/sell semantics in a single, explicit table.

corporate/SignatureStatus and SignatureCancel now build the URI
through url.URL{Path, RawQuery: url.Values.Encode()}.RequestURI()
instead of "?"+url.QueryEscape, which is more robust against a
baseURL that already carries query parameters.

monobank.Client.SetBaseURL records a parse error in optErr (surfaced
from the first Do) instead of silently keeping the old value — the
old behaviour meant a fat-finger in a staging config sent every
request to production.

+++ acquiring/types.go
@@ -254,10 +254,12 @@ type WalletData struct {
 	Status    WalletStatus `json:"status"`
 }
 
-// TipsInfo is the tip amount/recipient on an invoice.
+// TipsInfo is the tip amount/recipient on an invoice. Amount is in
+// minor units of the invoice's currency (kopecks for UAH). int64
+// avoids the 32-bit-platform overflow trap on the bare int type.
 type TipsInfo struct {
 	EmployeeID string `json:"employeeId"`
-	Amount     int    `json:"amount,omitempty"`
+	Amount     int64  `json:"amount,omitempty"`
 }
 
 // ProcessingStatus is the state of an asynchronous operation
